+++ src/synapse/static/StaticData.cpp
@@ -19,7 +19,7 @@ void *mallocStatic()
 	return (void*)p;
 }
 
-int allocStaticPara(void *pCPU, int num)
+int allocStaticPara(void *pCPU, size_t num)
 {
 	StaticData *p = (StaticData*)pCPU;
 
@@ -34,7 +34,7 @@ int allocStaticPara(void *pCPU, int num)
 	return 0;
 }
 
-void *allocStatic(int num)
+void *allocStatic(size_t num)
 {
 	void *p = mallocStatic();
 	allocStaticPara(p, num);
@@ -64,7 +64,7 @@ int freeStatic(void *pCPU)
 	return 0;
 }
 
-int saveStatic(void *pCPU, int num, FILE *f)
+int saveStatic(void *pCPU, size_t num, FILE *f)
 {
 
 	StaticData *p = (StaticData*)pCPU;
@@ -73,19 +73,19 @@ int saveStatic(void *pCPU, int num, FILE *f)
 		num = p->num;
 	}
 
-	fwrite(&(num), sizeof(int), 1, f);
+	fwrite(&(num), sizeof(size_t), 1, f);
 	// fwrite(p->pDst, sizeof(int), num, f);
 	fwrite(p->pWeight, sizeof(real), num, f);
 
 	return 0;
 }
 
-void *loadStatic(int num, FILE *f)
+void *loadStatic(size_t num, FILE *f)
 {
 	StaticData *p = (StaticData*)allocStatic(num);
 
 
-	fread(&(p->num), sizeof(int), 1, f);
+	fread(&(p->num), sizeof(size_t), 1, f);
 	assert(num == p->num);
 
 	// fread(p->pDst, sizeof(int), num, f);
@@ -94,7 +94,7 @@ void *loadStatic(int num, FILE *f)
 	return p;
 }
 
-bool isEqualStatic(void *p1, void *p2, int num)
+bool isEqualStatic(void *p1, void *p2, size_t num)
 {
 	StaticData *t1 = (StaticData*)p1;
 	StaticData *t2 = (StaticData*)p2;
